backend/pyproject.toml
[project]
name = "my-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncmy>=0.2.10",
    "fastapi[standard]>=0.121.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]

